Include tracking in backup creation

+++ server/src/main/kotlin/suwayomi/tachidesk/manga/impl/backup/models/Track.kt
@@ -11,7 +11,7 @@ interface Track : Serializable {
 
     var sync_id: Int
 
-    var media_id: Int
+    var media_id: Long
 
     var library_id: Long?
